Enhanced Domains
Change site query to account for enhanced domains. When enhanced domain is enabled, "Subdomain" field on "Site" object is null. So we need to change to change the query to check for status='Active'.

In SiteDetail, "/" is added at the end of the URL but it's not added in old site. So we need to remove that